requirements.txt
openai
pytrends
python-dotenv
pandas
streamlit